# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMind MVP 0.3 API contract",
    "version": "0.3.0",
    "description": "MVP 0.1/0.2已实现接口及MVP 0.3拟实现契约；新增路径仅为T004设计，运行端点由T012–T028交付。JSON是YAML 1.2子集。"
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/animation-media/{mediaId}/mp4": {"get": {"tags": ["Animation"],"summary": "读取或下载已审核MP4","operationId": "downloadAnimationMP4","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/MediaId"},{"$ref": "#/components/parameters/DownloadMedia"}],"responses": {"200": {"description": "仅owner已绑定且已审核的媒体；固定Content-Type、nosniff、private缓存、安全文件名与来源/版本响应头；默认inline、显式下载attachment。MP4支持Range。","content": {"video/mp4": {"schema": {"type": "string","format": "binary"}}},"headers": {"Content-Disposition": {"description": "默认inline供播放/字幕；download=true用attachment；均使用服务端安全文件名。","schema": {"type": "string"}},"X-Content-Type-Options": {"schema": {"type": "string","const": "nosniff"}},"Cache-Control": {"schema": {"type": "string","const": "private, no-store"}},"X-Animation-Media-ID": {"schema": {"$ref": "#/components/schemas/Identifier"}},"X-Animation-Template-ID": {"schema": {"type": "string"}},"X-Animation-Template-Version": {"schema": {"type": "string"}},"X-Animation-Subtitle-Version": {"schema": {"type": "string"}},"X-Animation-Review-Version": {"schema": {"type": "string"}},"X-Animation-Content-SHA256": {"description": "完整文件的内容摘要；206 Range仍指完整资源。","schema": {"type": "string","pattern": "^[0-9a-f]{64}$"}},"X-Animation-Generated-By": {"schema": {"type": "string","const": "reviewed-manim-template"}},"X-Animation-Published-At": {"schema": {"type": "string","format": "date-time"}}}},"206": {"description": "经授权的MP4部分内容。","content": {"video/mp4": {"schema": {"type": "string","format": "binary"}}},"headers": {"Content-Disposition": {"description": "默认inline供播放/字幕；download=true用attachment；均使用服务端安全文件名。","schema": {"type": "string"}},"X-Content-Type-Options": {"schema": {"type": "string","const": "nosniff"}},"Cache-Control": {"schema": {"type": "string","const": "private, no-store"}},"X-Animation-Media-ID": {"schema": {"$ref": "#/components/schemas/Identifier"}},"X-Animation-Template-ID": {"schema": {"type": "string"}},"X-Animation-Template-Version": {"schema": {"type": "string"}},"X-Animation-Subtitle-Version": {"schema": {"type": "string"}},"X-Animation-Review-Version": {"schema": {"type": "string"}},"X-Animation-Content-SHA256": {"description": "完整文件的内容摘要；206 Range仍指完整资源。","schema": {"type": "string","pattern": "^[0-9a-f]{64}$"}},"X-Animation-Generated-By": {"schema": {"type": "string","const": "reviewed-manim-template"}},"X-Animation-Published-At": {"schema": {"type": "string","format": "date-time"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"$ref": "#/components/responses/MediaNotReady"},"416": {"description": "无效或越界单段Range；code=RANGE_NOT_SATISFIABLE，返回Content-Range: bytes */size。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},"503": {"$ref": "#/components/responses/MediaUnavailable"}}}},
    "/api/animation-media/{mediaId}/srt": {"get": {"tags": ["Animation"],"summary": "读取或下载已审核SRT","operationId": "downloadAnimationSRT","security": [{"sessionCookie": []}],"parameters": [{"$ref": "#/components/parameters/MediaId"},{"$ref": "#/components/parameters/DownloadMedia"}],"responses": {"200": {"description": "仅owner已绑定且已审核的媒体；固定Content-Type、nosniff、private缓存、安全文件名与来源/版本响应头；默认inline、显式下载attachment。SRT使用UTF-8。","content": {"application/x-subrip": {"schema": {"type": "string","format": "binary"}}},"headers": {"Content-Disposition": {"description": "默认inline供播放/字幕；download=true用attachment；均使用服务端安全文件名。","schema": {"type": "string"}},"X-Content-Type-Options": {"schema": {"type": "string","const": "nosniff"}},"Cache-Control": {"schema": {"type": "string","const": "private, no-store"}},"X-Animation-Media-ID": {"schema": {"$ref": "#/components/schemas/Identifier"}},"X-Animation-Template-ID": {"schema": {"type": "string"}},"X-Animation-Template-Version": {"schema": {"type": "string"}},"X-Animation-Subtitle-Version": {"schema": {"type": "string"}},"X-Animation-Review-Version": {"schema": {"type": "string"}},"X-Animation-Content-SHA256": {"description": "完整文件的内容摘要；206 Range仍指完整资源。","schema": {"type": "string","pattern": "^[0-9a-f]{64}$"}},"X-Animation-Generated-By": {"schema": {"type": "string","const": "reviewed-manim-template"}},"X-Animation-Published-At": {"schema": {"type": "string","format": "date-time"}}}},"401": {"$ref": "#/components/responses/Unauthorized"},"404": {"$ref": "#/components/responses/NotFound"},"409": {"$ref": "#/components/responses/MediaNotReady"},"503": {"$ref": "#/components/responses/MediaUnavailable"}}}},
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "edumind_session", "description": "Secure、HttpOnly、SameSite=Lax、Host-only Cookie，2h 绝对有效期；数据库只存令牌哈希。"}},
    "parameters": {
...
      "MediaId": {"name": "mediaId","in": "path","required": true,"schema": {"$ref": "#/components/schemas/Identifier"}},
...
      "DownloadMedia": {"name": "download","in": "query","required": false,"schema": {"type": "boolean","default": false},"description": "false为播放器/字幕inline；true为附件下载，文件名仍由服务端生成。"}
    },
    "responses": {
...
      "Unauthorized": {"description": "会话缺失、失效或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "NotFound": {"description": "资源不存在，或不属于当前用户。", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "MediaNotReady": {"description": "未生成或尚未审核发布；code=MEDIA_NOT_READY。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "MediaUnavailable": {"description": "已引用媒体不可读取；code=MEDIA_UNAVAILABLE；客户端提示用户重新请求动画以恢复，不在下载请求中渲染。","content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
...
      "Identifier": {"type": "string", "pattern": "^[A-Za-z0-9_-]{8,64}$"},
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "CSRF_FAILED",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "IDEMPOTENCY_CONFLICT",
              "PROFILE_VERSION_CONFLICT",
              "PATH_VERSION_CONFLICT",
              "GRAPH_NODE_NOT_FOUND",
              "QUIZ_SUBMISSION_INVALID",
              "PATH_UNREACHABLE",
              "ACCOUNT_CONFLICT",
              "RATE_LIMITED",
              "PROVIDER_UNAVAILABLE",
              "REVIEW_REJECTED",
              "OPERATION_CANCELED",
              "JOB_STATE_CONFLICT",
              "RENDER_UNAVAILABLE",
              "RENDER_FAILED",
              "RENDER_INTERRUPTED",
              "MEDIA_UNAVAILABLE",
              "MEDIA_NOT_READY",
              "CLASSROOM_VERSION_CONFLICT",
              "SCENE_VERSION_CONFLICT",
              "EVENT_CURSOR_INVALID",
              "EVENT_CURSOR_EXPIRED",
              "MESSAGE_CURSOR_INVALID",
              "MESSAGE_CURSOR_EXPIRED",
              "REVIEW_UNAVAILABLE",
              "CONTENT_RETRACTED",
              "ROLE_DISABLED",
              "DEBATE_UNAVAILABLE",
              "DEBATE_ALREADY_ACTIVE",
              "DEBATE_NOT_ACTIVE",
              "DEBATE_CONTEXT_UNAVAILABLE",
              "DEBATE_PUBLICATION_FAILED",
              "DEBATE_INVALID_OUTPUT",
              "DEBATE_INTERRUPTED"
            ]
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "operation_id": {
            "$ref": "#/components/schemas/Identifier"
          }
        }
      },
...
    }
  },
...
}